\name{opossom.tissues}
\alias{opossom.tissues}
\title{Example data set.}
\description{A data set comprising of 12 selected human tissues.}
\usage{
data(opossom.tissues)
}
\format{
The data set is stored in RData (binary) format.
}
\details{
The data set was downloaded from Gene Expression Omnibus repository
(http://www.ncbi.nlm.nih.gov/geo, GEO accession no. GSE7307).
About 38,000 genes in more than 650 samples were measured using the Affymetrix
HGU133-Plus2 microarray. A subset of 12 selected tissues from different
categories is used as example data set for the oposSOM-package.
ENSEMBL-IDs are used as row-identifier to enable geneset overrepresentation analysis.
}
\source{
http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE7307
}

